Namespace shift, and removed orphaned code
